fix size_t

+++ novokhatskiy.danil/P4/searchMax.cpp
@@ -4,10 +4,10 @@ int novokhatskiy::searchMax(int matrix[], size_t rows, size_t cols)
 {
   int max = 0;
   size_t size = std::min(rows,cols);
-  for (int i = 0; i < size - 1; i++)
+  for (size_t i = 0; i < size - 1; i++)
   {
     int sum = 0;
-    for (int j = 0; j <= i * cols; j += cols -1)
+    for (size_t j = 0; j <= i * cols; j += cols -1)
     {
       sum += matrix[j];
     }
@@ -16,10 +16,10 @@ int novokhatskiy::searchMax(int matrix[], size_t rows, size_t cols)
       max = sum;
     }
   }
-  for (int i = size - 1 + cols; i < rows * cols; i += cols)
+  for (size_t i = size - 1 + cols; i < rows * cols; i += cols)
   {
     int sum = 0;
-    for (int j = i; j < rows * cols; j += cols - 1)
+    for (size_t j = i; j < rows * cols; j += cols - 1)
     {
       sum += matrix[j];
     }
